refactor(engine): relocate image/texture record family to DrawBatcher (T9e)
Fifth sub-PR of T9. Moves the 7 image/texture record methods (texture,
draw_image, draw_image_repeat, draw_image_nine_slice, draw_image_filtered,
draw_atlas, draw_texture) into the new batches/images.rs, byte-identical thin
painter shims. painter.rs 5064 -> 4511.

Seam adds a GpuResources dimension: texture/draw_texture borrow
&ExternalTextureRegistry (read), draw_image/draw_atlas/draw_image_repeat/
draw_image_nine_slice borrow &mut TextureCache (cache load), all disjoint from
&mut current_segment + &GpuStateStack. draw_image_filtered is the outlier — it
threads &mut draw_order + opacity because its ColorFilter::Mode branch calls the
batcher rect; the painter shim reads compositor.current_opacity() once and
passes it (no double-read). texture_batch is NOT touched (flush-side, stays for
T10). Recursive repeat/nine_slice -> draw_image delegation and TextureId cache
keying preserved byte-identical.

Gates: fmt/taplo/clippy(both modes)/doc(both feature sets)/typos/port-check
green. GPU-readback serial (DX12) 211/0/7 — added
draw_image_filtered_mode_overlays_tint (transparent source isolates the
opacity-threaded Mode overlay; fails if the opacity thread drops or doubles).

Reviewed by rust-reviewer + ce-kieran (both COMPLETE; all 7 bodies byte-identical
vs main, resources borrow regime, nine-slice 9-region decomposition, filtered
opacity-thread + z-order, scissor timing verified).

Found (deferred, NOT fixed here to keep the relocation byte-identical): the
ColorFilter::Mode tint rect is flushed UNDER the image (segment flush order is
rect_batch before cached_images at painter.rs ~560/563), so the Mode filter has
no visible effect on an OPAQUE image. Pre-existing; flagged for a follow-up with
a .flutter/ cross-check + a readback test that fails on today's code.
